// packages/services/shopify/src/__generated__/graphql.ts
export type BaseCartLineConnection = {
  readonly __typename: 'BaseCartLineConnection';
  /** A list of edges. */
  readonly edges: ReadonlyArray<BaseCartLineEdge>;
  /** A list of the nodes contained in BaseCartLineEdge. */
  readonly nodes: ReadonlyArray<BaseCartLine>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
};

/**
 * An auto-generated type which holds one BaseCartLine and a cursor during pagination.
 *
 */
export type BaseCartLineEdge = {
  readonly __typename: 'BaseCartLineEdge';
  /** A cursor for use in pagination. */
  readonly cursor: Scalars['String']['output'];
  /** The item at the end of BaseCartLineEdge. */
  readonly node: BaseCartLine;
};
// ...
export type Cart = HasMetafields & Node & {
  readonly __typename: 'Cart';
  /** The gift cards that have been applied to the cart. */
  readonly appliedGiftCards: ReadonlyArray<AppliedGiftCard>;
  /** An attribute associated with the cart. */
  readonly attribute: Maybe<Attribute>;
  /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
  readonly attributes: ReadonlyArray<Attribute>;
  /** Information about the buyer that's interacting with the cart. */
  readonly buyerIdentity: CartBuyerIdentity;
  /** The URL of the checkout for the cart. */
  readonly checkoutUrl: Scalars['URL']['output'];
  /** The estimated costs that the buyer will pay at checkout. The costs are subject to change and changes will be reflected at checkout. The `cost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing). */
  readonly cost: CartCost;
  /** The date and time when the cart was created. */
  readonly createdAt: Scalars['DateTime']['output'];
  /** The delivery properties of the cart. */
  readonly delivery: CartDelivery;
  /**
   * The delivery groups available for the cart, based on the buyer identity default
   * delivery address preference or the default address of the logged-in customer.
   *
   */
  readonly deliveryGroups: CartDeliveryGroupConnection;
  /** The discounts that have been applied to the entire cart. */
  readonly discountAllocations: ReadonlyArray<CartDiscountAllocation>;
  /** The case-insensitive discount codes that the customer added at checkout. */
  readonly discountCodes: ReadonlyArray<CartDiscountCode>;
  /**
   * The estimated costs that the buyer will pay at checkout. The estimated costs are subject to change and changes will be reflected at checkout. The `estimatedCost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing).
   * @deprecated Use `cost` instead.
   */
  readonly estimatedCost: CartEstimatedCost;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** A list of lines containing information about the items the customer intends to purchase. */
  readonly lines: BaseCartLineConnection;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
  /** A note that's associated with the cart. For example, the note can be a personalized message to the buyer. */
  readonly note: Maybe<Scalars['String']['output']>;
  /** The total number of items in the cart. */
  readonly totalQuantity: Scalars['Int']['output'];
  /** The date and time when the cart was updated. */
  readonly updatedAt: Scalars['DateTime']['output'];
};
// ...
export type CartLine = BaseCartLine & Node & {
  readonly __typename: 'CartLine';
  /** An attribute associated with the cart line. */
  readonly attribute: Maybe<Attribute>;
  /** The attributes associated with the cart line. Attributes are represented as key-value pairs. */
  readonly attributes: ReadonlyArray<Attribute>;
  /** The cost of the merchandise that the buyer will pay for at checkout. The costs are subject to change and changes will be reflected at checkout. */
  readonly cost: CartLineCost;
  /** The discounts that have been applied to the cart line. */
  readonly discountAllocations: ReadonlyArray<CartDiscountAllocation>;
  /**
   * The estimated cost of the merchandise that the buyer will pay for at checkout. The estimated costs are subject to change and changes will be reflected at checkout.
   * @deprecated Use `cost` instead.
   */
  readonly estimatedCost: CartLineEstimatedCost;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The merchandise that the buyer intends to purchase. */
  readonly merchandise: Merchandise;
  /** The quantity of the merchandise that the customer intends to purchase. */
  readonly quantity: Scalars['Int']['output'];
  /** The selling plan associated with the cart line and the effect that each selling plan has on variants when they're purchased. */
  readonly sellingPlanAllocation: Maybe<SellingPlanAllocation>;
};
// ...
export type HasMetafields = {
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
};
// ...
export type MoneyV2 = {
  readonly __typename: 'MoneyV2';
  /** Decimal money amount. */
  readonly amount: Scalars['Decimal']['output'];
  /** Currency of the money. */
  readonly currencyCode: CurrencyCode;
};
// ...
export type ProductVariant = HasMetafields & Node & {
  readonly __typename: 'ProductVariant';
  /** Indicates if the product variant is available for sale. */
  readonly availableForSale: Scalars['Boolean']['output'];
  /** The barcode (for example, ISBN, UPC, or GTIN) associated with the variant. */
  readonly barcode: Maybe<Scalars['String']['output']>;
  /** The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPrice` is higher than `price`. */
  readonly compareAtPrice: Maybe<MoneyV2>;
  /**
   * The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPriceV2` is higher than `priceV2`.
   * @deprecated Use `compareAtPrice` instead.
   */
  readonly compareAtPriceV2: Maybe<MoneyV2>;
  /**
   * List of bundles components included in the variant considering only fixed bundles.
   *
   */
  readonly components: ProductVariantComponentConnection;
  /** Whether a product is out of stock but still available for purchase (used for backorders). */
  readonly currentlyNotInStock: Scalars['Boolean']['output'];
  /**
   * List of bundles that include this variant considering only fixed bundles.
   *
   */
  readonly groupedBy: ProductVariantConnection;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** Image associated with the product variant. This field falls back to the product image if no image is available. */
  readonly image: Maybe<Image>;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
  /** The product variant’s price. */
  readonly price: MoneyV2;
  /**
   * The product variant’s price.
   * @deprecated Use `price` instead.
   */
  readonly priceV2: MoneyV2;
  /** The product object that the product variant belongs to. */
  readonly product: Product;
  /** The total sellable quantity of the variant for online sales channels. */
  readonly quantityAvailable: Maybe<Scalars['Int']['output']>;
  /** A list of quantity breaks for the product variant. */
  readonly quantityPriceBreaks: QuantityPriceBreakConnection;
  /** The quantity rule for the product variant in a given context. */
  readonly quantityRule: QuantityRule;
  /**
   * Whether a product variant requires components. The default value is `false`.
   * If `true`, then the product variant can only be purchased as a parent bundle with components.
   *
   */
  readonly requiresComponents: Scalars['Boolean']['output'];
  /** Whether a customer needs to provide a shipping address when placing an order for the product variant. */
  readonly requiresShipping: Scalars['Boolean']['output'];
  /** List of product options applied to the variant. */
  readonly selectedOptions: ReadonlyArray<SelectedOption>;
  /** Represents an association between a variant and a selling plan. Selling plan allocations describe which selling plans are available for each variant, and what their impact is on pricing. */
  readonly sellingPlanAllocations: SellingPlanAllocationConnection;
  /** The Shop Pay Installments pricing information for the product variant. */
  readonly shopPayInstallmentsPricing: Maybe<ShopPayInstallmentsProductVariantPricing>;
  /** The SKU (stock keeping unit) associated with the variant. */
  readonly sku: Maybe<Scalars['String']['output']>;
  /** The in-store pickup availability of this variant by location. */
  readonly storeAvailability: StoreAvailabilityConnection;
  /** Whether tax is charged when the product variant is sold. */
  readonly taxable: Scalars['Boolean']['output'];
  /** The product variant’s title. */
  readonly title: Scalars['String']['output'];
  /** The unit price value for the variant based on the variant's measurement. */
  readonly unitPrice: Maybe<MoneyV2>;
  /** The unit price measurement for the variant. */
  readonly unitPriceMeasurement: Maybe<UnitPriceMeasurement>;
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  readonly weight: Maybe<Scalars['Float']['output']>;
  /** Unit of measurement for weight. */
  readonly weightUnit: WeightUnit;
};
// ...
export type CartDetailedFragment = { readonly __typename: 'Cart', readonly checkoutUrl: any, readonly id: string, readonly totalQuantity: number, readonly lines: { readonly __typename: 'BaseCartLineConnection', readonly edges: ReadonlyArray<{ readonly __typename: 'BaseCartLineEdge', readonly node: { readonly __typename: 'CartLine', readonly id: string, readonly quantity: number, readonly merchandise: { readonly __typename: 'ProductVariant', readonly id: string, readonly title: string, readonly availableForSale: boolean, readonly quantityAvailable: number | null, readonly price: { readonly __typename: 'MoneyV2', readonly amount: any, readonly currencyCode: CurrencyCode }, readonly image: { readonly __typename: 'Image', readonly url: any } | null, readonly product: { readonly __typename: 'Product', readonly title: string, readonly featuredImage: { readonly __typename: 'Image', readonly url: any } | null } } } | { readonly __typename: 'ComponentizableCartLine', readonly id: string, readonly quantity: number, readonly merchandise: { readonly __typename: 'ProductVariant', readonly id: string, readonly title: string, readonly availableForSale: boolean, readonly quantityAvailable: number | null, readonly price: { readonly __typename: 'MoneyV2', readonly amount: any, readonly currencyCode: CurrencyCode }, readonly image: { readonly __typename: 'Image', readonly url: any } | null, readonly product: { readonly __typename: 'Product', readonly title: string, readonly featuredImage: { readonly __typename: 'Image', readonly url: any } | null } } } }> }, readonly cost: { readonly __typename: 'CartCost', readonly totalAmount: { readonly __typename: 'MoneyV2', readonly amount: any, readonly currencyCode: CurrencyCode } } };